Move appropriate classes to controller
