+++ app/src/main/java/com/example/botb/ConnectionActivity.java
@@ -5,6 +5,8 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.widget.TextView;
+import com.example.botb.controller.InputManager;
+import com.example.botb.controller.InputSubscriber;
 
 public class ConnectionActivity extends AppCompatActivity implements InputSubscriber {
