+++ app/src/main/java/com/example/botb/controller/ConnectionHandler.java
@@ -1,4 +1,4 @@
-package com.example.botb;
+package com.example.botb.controller;
 
 import android.util.Log;
 import java.io.IOException;
